Troubleshoot overflow by changing integer type to long type in payroll calculation logic

+++ src/main/kotlin/com/moa/service/dto/MonthlyEarningsResponse.kt
@@ -1,8 +1,8 @@
 package com.moa.service.dto
 
 data class MonthlyEarningsResponse(
-    val workedEarnings: Int,
-    val standardSalary: Int,
+    val workedEarnings: Long,
+    val standardSalary: Long,
     val workedMinutes: Long,
     val standardMinutes: Long,
 )
